+++ kll/utils.go
@@ -55,16 +55,16 @@ func getNaturalRank(normalizedRank float64, totalN uint64, inclusive bool) int64
 	return int64(math.Floor(naturalRank))
 }
 
-func checkK(k int, m int) error {
-	if k < m || k > _MAX_K {
-		return errors.New("K must be >= " + strconv.Itoa(m) + " and <= " + strconv.Itoa(_MAX_K) + ": " + strconv.Itoa(k))
+func checkK(k uint16, m uint8) error {
+	if k < uint16(m) || k > _MAX_K {
+		return errors.New("K must be >= " + strconv.Itoa(int(m)) + " and <= " + strconv.Itoa(_MAX_K) + ": " + strconv.Itoa(int(k)))
 	}
 	return nil
 }
 
-func checkM(m int) error {
+func checkM(m uint8) error {
 	if m < _MIN_M || m > _MAX_M || ((m & 1) == 1) {
-		return errors.New("M must be >= 2, <= 8 and even: " + strconv.Itoa(m))
+		return errors.New("M must be >= 2, <= 8 and even: " + strconv.Itoa(int(m)))
 	}
 	return nil
 }
